+++ src/config/index.ts
@@ -2,5 +2,5 @@ import { env } from "~/env";
 
 export const isLocal = env.VITE_APP_ENV === "local";
 
-export * from "./common.types";
-export * from "./storybook.types";
+export * from "./common";
+export * from "./storybook";
