--- src/config/providers/index.ts
@@ -1,3 +1,0 @@
-export * from "./Providers";
-export * from "./ReactQueryProvider";
-export * from "./SentryProvider";
