+++ src/providers/Providers.tsx
@@ -2,7 +2,7 @@ import type { PropsWithChildren } from "react";
 import React from "react";
 
 import { ReactQueryProvider } from "./ReactQueryProvider";
-import { SentryProvider } from "./SentryProvider";
+import { SentryProvider } from "./sentry/SentryProvider";
 
 export const Providers = ({ children }: PropsWithChildren) => {
   return (
