+++ src/shared/components/icons/CloseIcon.tsx
@@ -1,6 +1,6 @@
 import React from "react";
 
-import type { SVGProps } from "~/config/constants";
+import type { SVGProps } from "~/config";
 
 export const CloseIcon = ({ className, ...props }: SVGProps) => (
   <svg
