+++ tsconfig.json
@@ -11,7 +11,6 @@
     "paths": {
       "~/assets": ["./shared/assets"],
       "~/components": ["./shared/components/ui"],
-      "~/constants": ["./config/constants"],
       "~/hooks": ["./shared/hooks"],
       "~/icons": ["./shared/components/icons"],
       "~/sections": ["./shared/sections"],
